v0.1.4 ant procreation
ants procreate when exactly two fertile ants are on the same cell, and if three or more meet, they get destroyed. Added a fertility counter to render ants sterile for a certain number of steps after giving birth (the population boom was very exponential otherwise, out of control babies everywhere).