v0.2.0 multiple ants completed
completed the handling of multiple ants by adding visible labels for ticks and number of ants that update appropriately when game is run or when player manually changes the board.
Restructured a little the way signals are being sent and debugged an issue where the order of the ready functions in different parts of the tree would throw null exceptions.